Update auth.yaml
There was an error when I was doing a lab based on this file in Qwikilabs. 

The error was : 
error: unable to recognise “deployments/auth.yank”. no matches for kind “Deployment” in version “extensions/v1beta”

So I made changed in the apiVersion line 
and then added 3 lines in the sec: after replicas 1 
aiming at the proper selector.